requirements.txt
# Autonomous App Support AI Agent - Dependencies

# Core LangGraph & LangChain
langgraph>=0.3.27
langgraph-sdk>=0.1.66
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# Azure Services
azure-functions>=1.18.0
azure-servicebus>=7.11.0
azure-identity>=1.15.0
azure-monitor-query>=1.2.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0

# Databricks
databricks-sdk>=0.20.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pgvector>=0.2.4
asyncpg>=0.29.0

# SQL Server (for app databases)
pyodbc>=5.0.0

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tenacity>=8.2.0
structlog>=24.1.0

# API Framework
fastapi[standard]>=0.115.0
uvicorn>=0.27.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
